# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Portfolio APIS",
    "description": "Documentaion of Portfolio APIs",
    "contact": {
      "email": "sagar161994@gmai.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://portfolio-api-node.herokuapp.com"
    },
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/api/User/checkusername/{userName}": {
      "get": {
        "tags": [
          "SignUp"
        ],
        "summary": "Check Username",
        "description": "Check if username is already taken",
        "operationId": "Username",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "description": "Username needed",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "error",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}